refactor: migrated /datasets GET/POST, /datasets{id} GET/DELETE
